+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * disable `x-powered-by`
 * `config.db` moved to `config.model.db`
 * autoload model (`volebo-data`)
+* listening on local path (not host:port), handle existing socket correctly